// sdks/typescript/src/models/components/getinstanceresult.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod";
import { remap as remap$ } from "../../lib/primitives.js";
import { safeParse } from "../../lib/schemas.js";
import { Result as SafeParseResult } from "../../types/fp.js";
import { SDKValidationError } from "../errors/sdkvalidationerror.js";
import {
  Environment,
  Environment$inboundSchema,
  Environment$Outbound,
  Environment$outboundSchema,
} from "./environment.js";
import {
  HTTPToolDefinition,
  HTTPToolDefinition$inboundSchema,
  HTTPToolDefinition$Outbound,
  HTTPToolDefinition$outboundSchema,
} from "./httptooldefinition.js";

export type GetInstanceResult = {
  /**
   * The description of the toolset
   */
  description?: string | undefined;
  /**
   * Model representing an environment
   */
  environment: Environment;
  /**
   * The name of the toolset
   */
  name: string;
  /**
   * The environment variables that are relevant to the toolset
   */
  relevantEnvironmentVariables?: Array<string> | undefined;
  /**
   * The list of tools
   */
  tools: Array<HTTPToolDefinition>;
};

/** @internal */
export const GetInstanceResult$inboundSchema: z.ZodType<
  GetInstanceResult,
  z.ZodTypeDef,
  unknown
> = z.object({
  description: z.string().optional(),
  environment: Environment$inboundSchema,
  name: z.string(),
  relevant_environment_variables: z.array(z.string()).optional(),
  tools: z.array(HTTPToolDefinition$inboundSchema),
}).transform((v) => {
  return remap$(v, {
    "relevant_environment_variables": "relevantEnvironmentVariables",
  });
});

/** @internal */
export type GetInstanceResult$Outbound = {
  description?: string | undefined;
  environment: Environment$Outbound;
  name: string;
  relevant_environment_variables?: Array<string> | undefined;
  tools: Array<HTTPToolDefinition$Outbound>;
};

/** @internal */
export const GetInstanceResult$outboundSchema: z.ZodType<
  GetInstanceResult$Outbound,
  z.ZodTypeDef,
  GetInstanceResult
> = z.object({
  description: z.string().optional(),
  environment: Environment$outboundSchema,
  name: z.string(),
  relevantEnvironmentVariables: z.array(z.string()).optional(),
  tools: z.array(HTTPToolDefinition$outboundSchema),
}).transform((v) => {
  return remap$(v, {
    relevantEnvironmentVariables: "relevant_environment_variables",
  });
});

/**
 * @internal
 * @deprecated This namespace will be removed in future versions. Use schemas and types that are exported directly from this module.
 */
export namespace GetInstanceResult$ {
  /** @deprecated use `GetInstanceResult$inboundSchema` instead. */
  export const inboundSchema = GetInstanceResult$inboundSchema;
  /** @deprecated use `GetInstanceResult$outboundSchema` instead. */
  export const outboundSchema = GetInstanceResult$outboundSchema;
  /** @deprecated use `GetInstanceResult$Outbound` instead. */
  export type Outbound = GetInstanceResult$Outbound;
}

export function getInstanceResultToJSON(
  getInstanceResult: GetInstanceResult,
): string {
  return JSON.stringify(
    GetInstanceResult$outboundSchema.parse(getInstanceResult),
  );
}

export function getInstanceResultFromJSON(
  jsonString: string,
): SafeParseResult<GetInstanceResult, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => GetInstanceResult$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'GetInstanceResult' from JSON`,
  );
}